Move VatDetails type to packages/wpcom-checkout
