+++ client/me/purchases/vat-info/index.tsx
@@ -15,7 +15,8 @@ import { CALYPSO_CONTACT } from 'calypso/lib/url/support';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { errorNotice, successNotice, removeNotice } from 'calypso/state/notices/actions';
 import useVatDetails from './use-vat-details';
-import type { VatDetails, UpdateError } from './use-vat-details';
+import type { UpdateError } from './use-vat-details';
+import type { VatDetails } from '@automattic/wpcom-checkout';
 
 import './style.scss';
